docs/events/requirements.txt
pyyaml
genson
json-spec
check-jsonschema
